Revert "Fix build errors on dnxcore50"
This reverts commit b2c5bbffac4da5caf47ba1172f4161fff478acec.

+++ Enyim.Caching/Memcached/DefaultServerPool.cs
@@ -21,7 +21,7 @@ namespace Enyim.Caching.Memcached
 		private object DeadSync = new Object();
 		private System.Threading.Timer resurrectTimer;
 		private bool isTimerActive;
-		private int deadTimeoutMsec;
+		private long deadTimeoutMsec;
 		private bool isDisposed;
 		private event Action<IMemcachedNode> nodeFailed;
 
@@ -33,7 +33,7 @@ namespace Enyim.Caching.Memcached
 			this.configuration = configuration;
 			this.factory = opFactory;
 
-			this.deadTimeoutMsec = (int)this.configuration.SocketPool.DeadTimeout.TotalMilliseconds;
+			this.deadTimeoutMsec = (long)this.configuration.SocketPool.DeadTimeout.TotalMilliseconds;
 		}
 
 		~DefaultServerPool()
